backend/receptorium/requirements.txt
asgiref==3.3.4
Django==3.2
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
gunicorn==20.0.4
psycopg2-binary==2.8.6
PyJWT==2.1.0
pytz==2020.1
sqlparse==0.3.1 
flake8==4.0.1
djoser==2.1.0
webcolors==1.12
Pillow==9.2.0
drf-extra-fields==3.4.0